change architecture

+++ res/include/ud_math_array.h
@@ -12,7 +12,7 @@
 
 // Prototypes
 ud_arr*                             ud_matha_softmax(ud_arr* x);
-ud_arr                              *ud_matha_itoa(int n);
+ud_arr                              *ud_matha_itoa(long long n);
 float                               ud_matha_mean(ud_arr *x);
 float                               ud_matha_min_a(ud_arr *x);
 float                               ud_matha_max_a(ud_arr *x);
